move copy/pasted where() / orWhere() to WhereTrait
