+++ src/Common/Select.php
@@ -20,6 +20,8 @@ use Aura\SqlQuery\Exception;
  */
 class Select extends AbstractQuery implements SelectInterface, SubselectInterface
 {
+    use WhereTrait;
+
     /**
      *
      * An array of union SELECT statements.
@@ -977,44 +979,6 @@ class Select extends AbstractQuery implements SelectInterface, SubselectInterfac
         }
 
         return PHP_EOL . 'FOR UPDATE';
-    }
-
-    /**
-     *
-     * Adds a WHERE condition to the query by AND. If the condition has
-     * ?-placeholders, additional arguments to the method will be bound to
-     * those placeholders sequentially.
-     *
-     * @param string $cond The WHERE condition.
-     * @param mixed ...$bind arguments to bind to placeholders
-     *
-     * @return $this
-     *
-     */
-    public function where($cond, ...$bind)
-    {
-        $this->addWhere('AND', $cond, ...$bind);
-        return $this;
-    }
-
-    /**
-     *
-     * Adds a WHERE condition to the query by OR. If the condition has
-     * ?-placeholders, additional arguments to the method will be bound to
-     * those placeholders sequentially.
-     *
-     * @param string $cond The WHERE condition.
-     * @param mixed ...$bind arguments to bind to placeholders
-     *
-     * @return $this
-     *
-     * @see where()
-     *
-     */
-    public function orWhere($cond, ...$bind)
-    {
-        $this->addWhere('OR', $cond, ...$bind);
-        return $this;
     }
 
     /**
