+++ src/Common/Update.php
@@ -19,6 +19,8 @@ use Aura\SqlQuery\AbstractDmlQuery;
  */
 class Update extends AbstractDmlQuery implements UpdateInterface
 {
+    use WhereTrait;
+
     /**
      *
      * The table to update.
@@ -72,44 +74,6 @@ class Update extends AbstractDmlQuery implements UpdateInterface
     protected function buildTable()
     {
         return " {$this->table}";
-    }
-
-    /**
-     *
-     * Adds a WHERE condition to the query by AND. If the condition has
-     * ?-placeholders, additional arguments to the method will be bound to
-     * those placeholders sequentially.
-     *
-     * @param string $cond The WHERE condition.
-     * @param mixed ...$bind arguments to bind to placeholders
-     *
-     * @return $this
-     *
-     */
-    public function where($cond, ...$bind)
-    {
-        $this->addWhere('AND', $cond, ...$bind);
-        return $this;
-    }
-
-    /**
-     *
-     * Adds a WHERE condition to the query by OR. If the condition has
-     * ?-placeholders, additional arguments to the method will be bound to
-     * those placeholders sequentially.
-     *
-     * @param string $cond The WHERE condition.
-     * @param mixed ...$bind arguments to bind to placeholders
-     *
-     * @return $this
-     *
-     * @see where()
-     *
-     */
-    public function orWhere($cond, ...$bind)
-    {
-        $this->addWhere('OR', $cond, ...$bind);
-        return $this;
     }
 
     /**
